requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.3
python-decouple==3.8
psycopg2-binary==2.9.7
celery==5.3.4
redis==5.0.1
requests==2.31.0
python-jose[cryptography]==3.3.0
coverage==7.3.2
pytest==7.4.3
pytest-django==4.5.2
pytest-cov==4.1.0
gunicorn==21.2.0
Pillow==10.0.1
africastalking==1.2.5  # Added Africa's Talking SDK
pyjwt==2.8.0